# Speed up `cupy.vdot`