Close #3672. Use CUB device reduction to accelerate it. If not available, fall back to the original kernel.

Before: 
```
vdot                :    CPU:   17.272 us   +/- 0.938 (min:   16.300 / max:   39.416) us     GPU-0:  640.646 us   +/-18.809 (min:  630.496 / max:  774.144) us
```
After:
```
vdot                :    CPU:   57.839 us   +/- 6.615 (min:   55.680 / max:  216.692) us     GPU-0:   80.302 us   +/-18.042 (min:   75.776 / max:  232.096) us
```
For some reason CPU overhead is unexpectedly increased...Not sure what I overlooked 😞