@leofang 2020-08-04
I am still a bit concerned with the CPU overhead (57 us) which does not show up in my naive implementation (25 us).